+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **Breaking:** envs: Remove `get_bullet_action` function
 - **Breaking:** envs: Remove `set_bullet_action` function
+- cpp: Remove contact monitoring from Bullet interface
 - cpp: Remove external force processing from Bullet interface
 - cpp: Remove inertia randomization from Bullet interface
 - cpp: Remove joint friction from Bullet interface